A lot of cleanup, added a "ssh" attach feature
