+++ pide-rs/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "pide-rs"
-version = "0.2.0"
+version = "0.2.1"
 authors = ["James Munns <james.munns@gmail.com>"]
 description = "Persistent Isolated Development Environments (in Docker)"
 license = "MIT"
@@ -11,13 +11,14 @@ documentation = "https://docs.rs/pide-rs"
 [dependencies]
 clap = "2"
 log = "0.3"
-log4rs = "0.5"
 
 toml = "0.3"
 serde = "0.9"
 serde_derive = "0.9"
-chrono = { version = "0.3", features = ["serde"] }
 uuid = { version = "0.4", features = ["serde", "v4"] }
+
+[profile.release]
+lto = true
 
 [[bin]]
 name = "pide"
