+++ pide-rs/src/main.rs
@@ -4,12 +4,13 @@ extern crate clap;
 #[macro_use]
 extern crate serde_derive;
 
-extern crate chrono;
 extern crate serde;
 extern crate toml;
 extern crate uuid;
 
 mod cli;
+mod docker;
+mod types;
 
 fn main() {
     cli::parse_cli();
